Add new custom Error, small refactoring
