+++ errors/authRequired.js
@@ -1,0 +1,8 @@
+class authRequired extends Error {
+  constructor(message) {
+    super(message);
+    this.name = 'authRequired';
+    this.status = 403;
+  }
+}
+module.exports = authRequired;
